requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
python-dotenv
google-generativeai
transformers
torch
